fix: remove from

+++ kernel/src/main.rs
@@ -43,7 +43,7 @@ static NUM_READY_WORKER: AtomicU16 = AtomicU16::new(0);
 /// `Info` of `KernelInfo<Info>` represents architecture specific information.
 fn main<Info: Debug>(kernel_info: KernelInfo<Info>) {
     log::info!("CPU#{} is starting.", kernel_info.cpu_id);
-    IS_LOAD_RUNNING[usize::from(kernel_info.cpu_id)].store(true, Ordering::Relaxed);
+    IS_LOAD_RUNNING[kernel_info.cpu_id].store(true, Ordering::Relaxed);
 
     unsafe { awkernel_lib::cpu::increment_num_cpu() };
